refactor: extract shared message types to eliminate DRY violations
- Move slackMessage, messageSummary, messagesResponse, messagesResult
  to messages.go — these were duplicated across read_channel_messages.go
  and read_thread.go
- Extract toMessagesResult() helper to deduplicate the message mapping
  loop and cursor extraction that appeared in both Execute methods
- Remove per-action response/result types (readChannelMessagesResponse,
  readThreadResponse, readChannelMessagesResult, readThreadResult) in
  favor of shared types
- Verified no race conditions with -race flag
