+++ connectors/slack/read_thread_test.go
@@ -83,7 +83,7 @@ func TestReadThread_Success(t *testing.T) {
 		t.Fatalf("unexpected error: %v", err)
 	}
 
-	var data readThreadResult
+	var data messagesResult
 	if err := json.Unmarshal(result.Data, &data); err != nil {
 		t.Fatalf("failed to unmarshal result: %v", err)
 	}
